class based approach for mapathon #54
