--- API/mapathon/__init__.py
@@ -1,80 +1,0 @@
-# Copyright (C) 2021 Humanitarian OpenStreetmap Team
-
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Affero General Public License as
-# published by the Free Software Foundation, either version 3 of the
-# License, or (at your option) any later version.
-
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Affero General Public License for more details.
-
-# You should have received a copy of the GNU Affero General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-# Humanitarian OpenStreetmap Team
-# 1100 13th Street NW Suite 800 Washington, D.C. 20005
-# <info@hotosm.org>
-
-from typing import List, Union
-from pydantic import validator
-from .. import BaseModel
-from datetime import datetime, date, timedelta
-
-
-class MappedFeature(BaseModel):
-    feature: str
-    action: str
-    count: int
-
-
-class MapathonContributor(BaseModel):
-    user_id: int
-    username: str
-    total_buildings: int
-    mapped_tasks: int
-    validated_tasks: int
-    editors: str
-
-
-class MappedFeatureWithUser(MappedFeature):
-    username: str
-
-
-class MapathonSummary(BaseModel):
-    total_contributors: int
-    mapped_features: List[MappedFeature]
-
-
-class MapathonDetail(BaseModel):
-    mapped_features: List[MappedFeatureWithUser]
-    contributors: List[MapathonContributor]
-
-
-class MapathonRequestParams(BaseModel):
-    project_ids: List[int]
-    from_timestamp: Union[datetime, date]
-    to_timestamp: Union[datetime, date]
-    hashtags: List[str]
-
-    @validator("to_timestamp")
-    def check_timestamp_diffs(cls, value, values, **kwargs):
-        from_timestamp = values.get("from_timestamp")
-        timestamp_diff = value - from_timestamp
-        if timestamp_diff > timedelta(hours=24):
-            raise ValueError(
-                "Timestamp difference must be lower than 24 hours"
-            )
-
-        return value
-
-    @validator("hashtags")
-    def check_hashtag_filter(cls, value, values, **kwargs):
-        project_ids = values.get("project_ids")
-        if len(project_ids) == 0 and len(value) == 0:
-            raise ValueError(
-                "Empty lists found for both hashtags and project_ids params"
-            )
-
-        return value
